+++ cmd/titania.go/main.go
@@ -1,42 +1,30 @@
 package main
 
 import (
-	"bufio"
-	"encoding/json"
 	"os"
 
 	"github.com/cwd-k2/titania.go/pkg/tester"
 )
 
-const VERSION string = "v0.6.2"
+const VERSION string = "v0.7.0"
 
 func main() {
-	// ターゲットのディレクトリと言語，quiet
-	directories, languages, quiet := optparse()
+	directories := optparse()
 
 	tester.SetQuiet(quiet)
-	tester.SetLanguages(languages)
+	tester.SetLanguages(langs)
+	if tmpdir != "" {
+		tester.SetTmpDir(tmpdir)
+	}
 
-	outcomes := exec(directories)
+	uresults := exec(directories)
 
 	// 何もテストが実行されなかった場合
-	if len(outcomes) == 0 {
+	if len(uresults) == 0 {
 		println("There's no test in (sub)directory[ies].")
 		os.Exit(1)
 	}
 
-	final(outcomes)
-
-	// buffering
-	stdout := bufio.NewWriter(os.Stdout)
-	defer stdout.Flush()
-
-	enc := json.NewEncoder(stdout)
-	enc.SetIndent("", "  ")
-	enc.SetEscapeHTML(false)
-
-	// TODO: 全部メモリに持っておくのは辛いので (形式はそのままに) 分割して出力したい.
-	if err := enc.Encode(outcomes); err != nil {
-		panic(err)
-	}
+	final(uresults)
+	printjson(uresults)
 }