+++ examples/example_03/test_method.rb
@@ -1,0 +1,7 @@
+# output, input, answer can be separated by null charactor.
+output, _, answer = gets(nil).split("\0")
+
+# PASS and FAIL
+# and test method should output PASS or FAIL
+STDOUT.puts output == answer ? "FAIL" : "PASS"
+STDERR.puts "Just for fun, inversing the result."