+++ pkg/tester/test_method.go
@@ -1,31 +1,27 @@
 package tester
 
 import (
-	"io/ioutil"
 	"path/filepath"
-	"strings"
 
-	"github.com/cwd-k2/titania.go/internal/pkg/langtype"
+	"github.com/cwd-k2/titania.go/pkg/runner"
 )
 
 type TestMethod struct {
 	Name       string
 	Language   string
-	OnResult   string
-	SourceCode []byte
+	OnExit     int
+	FileName   string
 	InputOrder []string
 }
 
 type TestMethodConfig struct {
 	FileName   string   `json:"file_name"`
-	On         string   `json:"on"`          // one of SUCCESS, EXECUTION FAILURE, ...
+	OnExit     int      `json:"on_exit"`     // on_exit: 0, ...
 	InputOrder []string `json:"input_order"` // ex: [stdout, stderr, input, ...]
 }
 
-// Creates TestMethod struct.
 // if error occurred, then nil will be returned.
-// TODO: error handling.
-func NewTestMethod(basepath string, config TestMethodConfig) *TestMethod {
+func ReadTestMethod(basepath string, config TestMethodConfig) *TestMethod {
 	if config.FileName == "" {
 		return nil
 	}
@@ -37,23 +33,10 @@ func NewTestMethod(basepath string, config TestMethodConfig) *TestMethod {
 		name = filename
 	}
 
-	sourceCodeBS, err := ioutil.ReadFile(filename)
-	if err != nil {
-		logger.Printf("%+v\n", err)
-		return nil
-	}
-
-	language := langtype.LangType(filename)
+	language := runner.LangType(filename)
 	if language == "plain" {
 		logger.Println("Invalid test method.")
 		return nil
-	}
-
-	var onresult string
-	if len(config.On) != 0 {
-		onresult = strings.ToUpper(config.On)
-	} else {
-		onresult = "SUCCESS"
 	}
 
 	var inputorder []string
@@ -67,8 +50,8 @@ func NewTestMethod(basepath string, config TestMethodConfig) *TestMethod {
 	return &TestMethod{
 		Name:       name,
 		Language:   language,
-		OnResult:   onresult,
-		SourceCode: sourceCodeBS,
+		OnExit:     config.OnExit,
+		FileName:   filename,
 		InputOrder: inputorder,
 	}
 }